% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{gradDescentRData}
\alias{gradDescentRData}
\title{Data set of the package}
\description{
This is a dataset that collected experimentaly by Kennedy in 1954 to obtain the density value of CO2.
Parameter used in the experiment are temperature and pressure, which it can be a parameter to
obtain compressibility factor value of CO2.
}
\references{
Kennedy, G. C. (1954). Pressure-Volmue Temperature Relations in CO2 at Elevated Temperatures and Pressures.
American Journal of Science, 225-241.
}
\keyword{data}
